requirements.txt
defusedxml~=0.7.1
dotmap~=1.3.26
argon2-cffi~=21.3.0